move ConnectionStreamHandlerTests into separate file
